Module Geneweb.GWPARAM

type syslog_level = [

| `LOG_ALERT

| `LOG_CRIT

| `LOG_DEBUG

| `LOG_EMERG

| `LOG_ERR

| `LOG_INFO

| `LOG_NOTICE

| `LOG_WARNING

]

The level of log gravity (`LOG_WARNING the lightest).

val init : (unit -> unit) Stdlib.ref

Inititialise assets for gwd server (one in current directory one in /usr/share/geneweb)

val base_path : (string list -> string -> string) Stdlib.ref

!base_path pref fname default function that returns path to fname inside base directory where pref is a list of subdirectories between base directory and fname.

val bpath : (string -> string) Stdlib.ref

!bpath fname default function that returns path to fname inside base directory

val output_error : (?headers:string list -> ?content:string -> Config.config -> Def.httpStatus -> unit) Stdlib.ref

!output_error ?headers ?content conf status default function that send the http status status, headers and content if provided. Otherwise send default content from /etc/<status-code>-<lang>.html

val p_auth : (Config.config -> Gwdb.base -> Gwdb.person -> bool) Stdlib.ref

Calculate the access rights to the person's information in according to his age. Returns (in the order of the tests) :

True if : requester is wizard or friend or person is public

True if : person has at least one title and public_if_title is set to yes in gwf config file

False if : person is alive and private_years > 0

True if : person is older (depending on the date of birth or baptism date) then privates_years

False if : person is younger (depending on the date of birth or baptism date) then privates_years

True if : person has been deceased for more than privates_years

False if : person has been deceased for less than privates_years

True if : person is between 80 and 120 years old and he is not beeing private and public_if_no_date is set to yes in gwf config file

True if : person has been married for more than private_years

False otherwise

val syslog : (syslog_level -> string -> unit) Stdlib.ref

!syslog level log log message log with gravity level level on stderr.

val wrap_output : (Config.config -> string -> (unit -> unit) -> unit) Stdlib.ref