Module GWPARAM.Default

val init : unit -> unit

Inititialise assets directories for gwd server: * current directory

val base_path : string list -> string -> string

Use concatenation of Secure.base_dir (), pref and fname

val bpath : string -> string

Filename.concat (Secure.base_dir ())

val output_error : ?headers:string list -> ?content:Adef.safe_string -> Config.config -> Def.httpStatus -> unit

If ?content is not set, sends page content from /etc/<status-code>-<lang>.html. If the current lang is not available, use `en`

val p_auth : Config.config -> Gwdb.base -> Gwdb.person -> bool

Calculate the access rights to the person's information in according to his age. Returns (in the order of the tests) :

`true` if requester is wizard or friend or person is public

`true` if person has at least one title and public_if_title is set to yes in gwf config file

`false` if person is alive and private_years > 0

`true` if person is older (depending on the date of birth or baptism date) then privates_years

`false` if person is younger (depending on the date of birth or baptism date) then privates_years

`true` if person has been deceased for more than privates_years

`false` if person has been deceased for less than privates_years

`true` if person is between 80 and 120 years old and he is not beeing private and public_if_no_date is set to yes in gwf config file

`true` if person has been married for more than private_years

`false` otherwise

val syslog : syslog_level -> string -> unit

Prints on stderr using `"date: level message"` format.

val wrap_output : Config.config -> Adef.safe_string -> (unit -> unit) -> unit

Display in a very basic HTML doc, with no CSS or JavaScript.