Module Geneweb.HistoryDiffDisplay

val print_clean : Config.config -> unit

Displays page that allows to select all revision of the history file in argument that user may want to clean

val print_clean_ok : Config.config -> unit

Cleans the history associated to the history file in argument

val print : Config.config -> Gwdb.base -> unit

Displays the page that allows to select (with variable t = "SUM") and to view (with variable t = "DIFF") the difference between all revisions of history file of concerned person in variable f. Intepretate the template file updhist_diff.txt