Module Geneweb.Hutil

type 'a value =

| Vint of int

| Vother of 'a

| Vnone
(*
Polymorphic value type for template environment
*)

module BufferPool : sig ... end

module TemplateCache : sig ... end

module HtmlBuffer : sig ... end

val get_env : string -> 'a value Templ.Env.t -> 'a value

get_env key env retrieves value from template environment. Returns Vnone if key doesn't exist.

val get_vother : 'a value -> 'a option

get_vother v extracts value if v = Vother x, else None.

val set_vother : 'a -> 'a value

set_vother x wraps value in Vother constructor.

val header_without_http_nor_home : Config.config -> (bool -> unit) -> unit

Base header without HTTP headers or home template. The boolean parameter determines if title is printed in <title> tag.

val header_with_title : ?error:bool -> ?fluid:bool -> Config.config -> (bool -> unit) -> unit

Complete header with title.

parameter error
Display title in red if true (default: false)

parameter fluid
Use container-fluid if true (default: false)

parameter conf
Base configuration

parameter title
Title function (bool indicates <title> vs <h1>)

val header_with_adaptive_title : ?fluid:bool -> Config.config -> string -> unit

Header with adaptive title size based on content length.

parameter fluid
Use container-fluid if true (default: false)

parameter conf
Base configuration

parameter title_content
Title text content (HTML)

val header_fluid : Config.config -> (bool -> unit) -> unit

Shortcut for header_with_title ~fluid:true.

val header_with_conf_title : Config.config -> (bool -> unit) -> unit

Header that gets title from conf.env "p_title".

val header_without_title : Config.config -> unit

Complete header without <h1> title element.

val header_without_home : Config.config -> (bool -> unit) -> unit

Like header_with_title but without home.txt inclusion.

val header : ?error:bool -> ?fluid:bool -> Config.config -> (bool -> unit) -> unit

Main header interface. Alias for header_with_title.

parameter error
Display title in red if true

parameter fluid
Use container-fluid if true

val rheader : Config.config -> (bool -> unit) -> unit

Alias for header ~error:true. Shows error title (red).

val trailer : Config.config -> unit

Standard HTML page footer. Outputs in order:

"trl" template (user trailer)

"copyr" template (copyright)

Container closure

"js" template (JavaScript)

Debug timing if enabled

</body></html> tags

val trailer_with_extra_js : Config.config -> string -> unit

trailer_with_extra_js conf sources outputs trailer with additional JavaScript injection before main js.txt.

parameter sources
JavaScript sources to inject:

Single template: "template_name"

Multiple templates: "templ1|templ2|templ3"

Inline JavaScript: "<script>...</script>"

Templates are cached for performance.

val link_to_referer : Config.config -> Adef.safe_string

Returns HTML link to previous page (referer). Returns empty string if no referer available.

val incorrect_request : ?comment:string -> Config.config -> unit

Sends HTTP 400 Bad Request response.

parameter comment
Optional error message

val error_cannot_access : Config.config -> string -> unit

Sends HTTP 404 Not Found when template cannot be accessed.

parameter fname
Name of inaccessible file

val include_home_template : Config.config -> unit

Includes home.txt template if available. Silent if file doesn't exist.