Module Hutil.HtmlBuffer

val create_buffered_conf : ?initial_size:int -> Config.config -> Config.config

create_buffered_conf ~initial_size conf returns a config with buffered output. All writes accumulate in a buffer until flush is called.

parameter initial_size
Initial buffer size in bytes (default: 65536)

val wrap : Config.config -> (Config.config -> 'a) -> 'a

wrap ~initial_size conf f executes f with a buffered config, automatically flushing at the end or on exception. Uses page buffer pool for efficiency.

val wrap_measured : Config.config -> (Config.config -> 'a) -> 'a

wrap_measured conf f like wrap but collects performance metrics. If "debug_buffer" is in conf.env, prints statistics to stderr:

Number of writes

Total bytes written

Time elapsed

Memory allocated