Module Geneweb.Output

val status : Config.config -> Def.httpStatus -> unit

status conf answer print HTTP status line to the socket where answer is a HTTP status.

val header : Config.config -> ('a, unit, string, unit) Stdlib.format4 -> 'a

Formatter printing of the HTTP header (header line) to the socket. If used without seting HTTP status with status, it is set OK.

val print_string : Config.config -> string -> unit

Printing the part of HTTP response body on the socket. If used without seting HTTP status with status, it is set OK.

val printf : Config.config -> ('a, unit, string, unit) Stdlib.format4 -> 'a

Formatter printing of the part of HTTP response body on the socket. If used without seting HTTP status with status, it is set OK.

val flush : Config.config -> unit

Flushes (send) printed previously content of the buffer on the socket.