Module Geneweb.Perso

type generation_person =

| GP_person of Sosa.t * Gwdb.iper * Gwdb.ifam option

| GP_same of Sosa.t * Sosa.t * Gwdb.iper

| GP_interv of (Sosa.t * Sosa.t * (Sosa.t * Sosa.t) option) option

| GP_missing of Sosa.t * Gwdb.iper

val string_of_marriage_text : Config.config -> Gwdb.base -> Gwdb.family -> Adef.safe_string

val interp_templ : ?no_headers:bool -> string -> Config.config -> Gwdb.base -> Gwdb.person -> unit

val interp_templ_with_menu : ( bool -> unit ) -> string -> Config.config -> Gwdb.base -> Gwdb.person -> unit

val interp_notempl_with_menu : ( bool -> unit ) -> string -> Config.config -> Gwdb.base -> Gwdb.person -> unit

val print : ?no_headers:bool -> Config.config -> Gwdb.base -> Gwdb.person -> unit

Displays the HTML page of a person

val print_ascend : Config.config -> Gwdb.base -> Gwdb.person -> unit

Displays the ascendants of the selected person

val print_what_links : Config.config -> Gwdb.base -> Gwdb.person -> unit

Displays links to pages associated to the person

val links_to_ind : Config.config -> Gwdb.base -> (( Gwdb.iper, Gwdb.ifam ) Def.NLDB.page * ('a * ((string * string * int) * 'b) list)) list -> (string * string * int) -> ( Gwdb.iper, Gwdb.ifam ) Def.NLDB.page list

val build_sosa_tree_ht : Config.config -> Gwdb.base -> Gwdb.person -> unit

Construts from the giving person sosa table strored in the cache. Sosa table contains association person_id -> sosa number for each person in the base. Person has sosa Sosa.one and his ancestors have sosa > Sosa.one. For non ancestor person sosa number is set to Sosa.zero.

val build_sosa_ht : Config.config -> Gwdb.base -> unit

Extract referenced person from environement and constructs for him sosa table wiht build_sosa_tree_ht.

val get_sosa_person : Gwdb.person -> Sosa.t

val get_single_sosa : Config.config -> Gwdb.base -> Gwdb.person -> Sosa.t

val print_sosa : Config.config -> Gwdb.base -> Gwdb.person -> bool -> unit

val get_linked_page : Config.config -> Gwdb.base -> Gwdb.person -> string -> Adef.safe_string

val get_birth_text : Config.config -> Gwdb.person -> bool -> Adef.safe_string

val get_baptism_text : Config.config -> Gwdb.person -> bool -> Adef.safe_string

val get_death_text : Config.config -> Gwdb.person -> bool -> Adef.safe_string

val get_burial_text : Config.config -> Gwdb.person -> bool -> Adef.safe_string

val get_cremation_text : Config.config -> Gwdb.person -> bool -> Adef.safe_string

val get_marriage_date_text : Config.config -> Gwdb.family -> bool -> Adef.safe_string

val linked_page_text : Config.config -> Gwdb.base -> Gwdb.person -> string -> 'a -> Adef.safe_string -> (( Gwdb.iper, Gwdb.ifam ) Def.NLDB.page * ('b * ('a * Def.NLDB.ind) list)) -> Adef.safe_string

val max_ancestor_level : Config.config -> Gwdb.base -> Gwdb.iper -> int -> int

val string_of_died : Config.config -> Gwdb.person -> bool -> Adef.safe_string

val string_of_parent_age : Config.config -> Gwdb.base -> (Gwdb.person * bool) -> ( Gwdb.family -> Gwdb.iper ) -> Adef.safe_string

val string_of_image_url : Config.config -> Gwdb.base -> (Gwdb.person * bool) -> bool -> Adef.escaped_string

val round_2_dec : float -> float

val has_children : Gwdb.base -> Gwdb.person -> bool

val string_of_image_size : Config.config -> Gwdb.base -> (Gwdb.person * bool) -> string

val string_of_image_medium_size : Config.config -> Gwdb.base -> (Gwdb.person * bool) -> string

val string_of_image_small_size : Config.config -> Gwdb.base -> (Gwdb.person * bool) -> string

val get_link : generation_person list -> Util.IperSet.elt -> generation_person option

val infinite : int

val limit_desc : Config.config -> int

val make_desc_level_table : Config.config -> Gwdb.base -> int -> Gwdb.person -> ( Util.IperSet.elt, int ) Gwdb.Marker.t * ( Gwdb.ifam, int ) Gwdb.Marker.t

val default_max_cousin_lev : int

type dup =

| DupFam of Gwdb.ifam * Gwdb.ifam

| DupInd of Gwdb.iper * Gwdb.iper

| NoDup

type excl_dup = (Gwdb.iper * Gwdb.iper) list * (Gwdb.ifam * Gwdb.ifam) list

val excluded_possible_duplications : Config.config -> excl_dup

val first_possible_duplication : Gwdb.base -> Gwdb.iper -> excl_dup -> dup

val nobility_titles_list : Config.config -> Gwdb.base -> Gwdb.person -> (int * Gwdb.istr Def.gen_title_name * Gwdb.istr * Gwdb.istr list * (Adef.date option * Adef.date option) list) list

val has_history : Config.config -> Gwdb.base -> Gwdb.person -> bool -> bool

val has_possible_duplications : Config.config -> Gwdb.base -> Gwdb.person -> bool

val string_of_title : ?link:bool -> Config.config -> Gwdb.base -> Adef.safe_string -> Gwdb.person -> (int * Gwdb.istr Def.gen_title_name * Gwdb.istr * Gwdb.istr list * (Def.date option * Def.date option) list) -> Adef.safe_string

Optionnal link argument is passed to DateDisplay.string_of_ondate

type event_name =

| Pevent of Gwdb.istr Def.gen_pers_event_name

| Fevent of Gwdb.istr Def.gen_fam_event_name

val events_list : Config.config -> Gwdb.base -> Gwdb.person -> (event_name * Def.cdate * Gwdb.istr * Gwdb.istr * Gwdb.istr * (Gwdb.iper * Def.witness_kind) array * Gwdb.iper option) list