Module Geneweb.UpdateData

module PersMap : sig ... end

module PersSet : sig ... end

module StringSet : sig ... end

module IstrSet : sig ... end

val get_data : Config.config -> (Gwdb.person -> Gwdb.istr) list * (('a, 'b) Def.gen_pers_event -> 'c) list * (Gwdb.family -> Gwdb.istr) list * (('d, 'e) Def.gen_fam_event -> 'f) list

val get_all_data : Config.config -> Gwdb.base -> IstrSet.elt list

val get_person_from_data : Config.config -> Gwdb.base -> (PersMap.key * PersSet.elt list) list

val combine_by_ini : string -> ('a * string) list -> (string * ('a * string) list) list

val reduce_cpl_list : int -> ('a * 'b list) list -> ('a * 'b list) list

Description : Retourne la sous liste telle que la somme des longueurs des ('b list) soit égale à size. Args :

size : la taille de la liste retournée

list : la liste originale Retour :

list : la nouvelle liste dont la somme des ('b list) est égale à size Rem : Non exporté en clair hors de ce module.

val update_person : Config.config -> Gwdb.base -> string -> string -> Gwdb.person -> (Gwdb.iper, Gwdb.iper, Gwdb.istr) Def.gen_person

Description : Met à jour le/les champ(s) de la personne. Args :

conf : configuration de la base

base : base de donnée

old : l'ancien contenu

new_input : le nouveau contenu

p : person Retour :

gen_person iper istr : gen_person avec les champs modifiés Rem : Non exporté en clair hors de ce module.

val update_family : Config.config -> Gwdb.base -> string -> string -> Gwdb.family -> (Gwdb.iper, Gwdb.ifam, Gwdb.istr) Def.gen_family

Description : Met à jour le/les champ(s) de la famille. Args :

conf : configuration de la base

base : base de donnée

old : l'ancien contenu

new_input : le nouveau contenu

fam : family Retour :

gen_family ifam istr : gen_family avec les champs modifiés Rem : Non exporté en clair hors de ce module.

val update_person_list : Config.config -> Gwdb.base -> string -> (string * Gwdb.person list) list -> int -> int -> int

Description : Args :

conf : configuration

base : base

new_input : le nouveau contenu

list : la liste des (clé, person list)

nb_pers : le nombre de personnes concernées par la mise à jour

max_updates = le nombre maximum de persons que l'on met à jour Retour :

unit Rem : Non exporté en clair hors de ce module.

val move_particle : Gwdb.base -> string -> string

val build_list : Config.config -> Gwdb.base -> (Gwdb.istr * string) list

Get all the data and filter them if "s" is defined in conf.env

val build_list_short : Config.config -> ('a * string) list -> StringSet.elt list

val build_list_long : Config.config -> (Gwdb.istr * string) list -> (string * (Gwdb.istr * string) list) list