Module Geneweb.Wiki

type wiki_info = {

wi_mode : string;

wi_file_path : string -> string;

wi_person_exists : (string * string * int) -> bool;

wi_always_show_link : bool;

}

val syntax_links : Config.config -> wiki_info -> string -> string

val html_of_tlsw : Config.config -> string -> string list

Parses a whole TLSW text to a list of strings

val html_with_summary_of_tlsw : Config.config -> wiki_info -> (bool * string * string) option -> string -> string

HTML displaying a table of content for a TLSW file

val extract_sub_part : string -> int -> string list

extract_sub_part tlsw i Extracts the `i`th first TLSW sections of `tlsw`

val split_title_and_text : string -> (string * string) list * string

The argument is expected to have the form "KEY=value\n"... This function calculates each Key/value pair and puts it in a list; except for the key TITLE, that is the second element of the returned tuple. If there is no title defined, checks if the first line is not empty and does not start with '=' nor contains '<' nor '', in which case it is choosen as a first line. Otherwise, the title is the empty string.

val print_sub_part : Config.config -> wiki_info -> bool -> string -> string -> int -> string list -> unit

Prints an exctracted sub part

val print_mod_view_page : Config.config -> bool -> string -> string -> ( bool -> unit ) -> (string * string) list -> string -> unit

Prints an editable part

val print_mod_ok : Config.config -> wiki_info -> ( string -> string option ) -> ( string option -> string ) -> ( string -> (string * string) list * string ) -> ( string -> string -> unit ) -> ( string -> string ) -> bool -> unit

Commits the changes of a page

val notes_aliases : Config.config -> (string * string) list

Reads the notes alias file (conf.base_env.notes_alias_file or base_path/notes.alias). File format is "KEY value\n...", returns the list of (KEY,value)

val map_notes : (string * string) list -> string -> string

Given an alias list, finds the corresponding alias for a given string