Module Geneweb_compat.Seq

val empty : 'a Stdlib.Seq.t

empty is the empty sequence. It has no elements. Its length is 0.

val take : int -> 'a Stdlib.Seq.t -> 'a Stdlib.Seq.t

take n xs is the sequence of the first n elements of xs.

If xs has fewer than n elements, then take n xs is equivalent to xs.

n must be nonnegative.

raises Invalid_argument
if n is negative.

val concat : 'a Stdlib.Seq.t Stdlib.Seq.t -> 'a Stdlib.Seq.t

If xss is a sequence of sequences, then concat xss is its concatenation.

If xss is the sequence xs0; xs1; ... then concat xss is the sequence xs0 @ xs1 @ ....

val equal : ('a -> 'b -> bool) -> 'a Stdlib.Seq.t -> 'b Stdlib.Seq.t -> bool

Provided the function eq defines an equality on elements, equal eq xs ys determines whether the sequences xs and ys are pointwise equal.

May not terminate if both of the sequences xs and ys are infinite.