Module Collection.Marker

Markers for elements inside a collection.

type ('k, 'v) t

Markers are way to annotate (add extra information to) elements of a t.

val make : ('k -> int) -> 'k collection -> 'a -> ('k, 'a) t

val dummy : 'a -> 'b -> ('a, 'b) t

dummy k v create a dummy collection with no element. k and v are only used for typing. Useful for placeholders or for typing purpose.

val get : ('k, 'v) t -> 'k -> 'v

get marker key Return the annotation associated to key.

val set : ('k, 'v) t -> 'k -> 'v -> unit

set marker key value Set value as annotation associated to key.