Module Geneweb_db.Dbdisk

type dsk_person = (int, int, int) Def.gen_person

Extended person's entry in the base

type dsk_ascend = int Def.gen_ascend

Person's ascendants entry in the base

type dsk_union = int Def.gen_union

Person's union entry in the base

type dsk_family = (int, int, int) Def.gen_family

Family's entry in the base

type dsk_couple = int Def.gen_couple

Family's couple entry in the base

type dsk_descend = int Def.gen_descend

Family's descendants entry in the base

type dsk_title = int Def.gen_title

Nobility title in the base

type 'a record_access = {

load_array : unit -> unit;

get : int -> 'a;

get_nopending : int -> 'a;

mutable len : int;

output_array : Stdlib.out_channel -> unit;

clear_array : unit -> unit;

}

Type that define the functions to use to access and manipulate with database arrays.

type string_person_index = {

find : int -> int list;

cursor : string -> int;

next : int -> int;

}

Data structure for optimised search throughout index by name (surname or first name). Considers also patched persons.

type visible_record_access = {

v_write : unit -> unit;

v_get : (dsk_person -> bool) -> int -> bool;

}

type perm =

| RDONLY

| RDRW

type base_data = {

persons : dsk_person record_access;

ascends : dsk_ascend record_access;

unions : dsk_union record_access;

visible : visible_record_access;

families : dsk_family record_access;

couples : dsk_couple record_access;

descends : dsk_descend record_access;

strings : string record_access;

particles_txt : string list;

particles : Re.re Stdlib.Lazy.t;

bnotes : Def.base_notes;

bdir : string;

perm : perm;

}

Data part of database

type base_func = {

person_of_key : string -> string -> int -> int option;

persons_of_name : string -> int list;

strings_of_sname : string -> int list;

strings_of_fname : string -> int list;

persons_of_surname : string_person_index;

persons_of_first_name : string_person_index;

patch_person : int -> dsk_person -> unit;

patch_ascend : int -> dsk_ascend -> unit;

patch_union : int -> dsk_union -> unit;

patch_family : int -> dsk_family -> unit;

patch_couple : int -> dsk_couple -> unit;

patch_descend : int -> dsk_descend -> unit;

patch_name : string -> int -> unit;

insert_string : string -> int;

replace_string : string -> string -> int;

commit_patches : unit -> unit;

commit_notes : string -> string -> unit;

commit_wiznotes : string -> string -> unit;

nb_of_real_persons : unit -> int;

iper_exists : int -> bool;

ifam_exists : int -> bool;

}

Functionality part of database. Every modification of the base is stored in patches file. Note that, every modification firstly is pendent and should be commited to apply them and to update patches file with commit_patches.

type base_version =

| GnWb0020

| GnWb0021

| GnWb0022

| GnWb0023

| GnWb0024

Geneweb database version

type dsk_base = {

data : base_data;

func : base_func;

version : base_version;

}

Database representation: data and basic requests over this data.