Module Driver.Istr

type t = istr

include Indexed with type t := t

val dummy : t

val is_dummy : t -> bool

val hash : t -> int

Compute a hash of an indexed value. This function is just the identity because a t value is already a hash.

The hash should use only on values from a common database index.

val equal : t -> t -> bool

val compare : t -> t -> int

val to_string : t -> string

val of_string : string -> t

val pp : t Fmt.t

module Set : Stdlib.Set.S with type elt = t

module Map : Stdlib.Map.S with type key = t

module Table : Stdlib.Hashtbl.S with type key = t

val empty : t

Identifier of the empty string ("").

val quest : t

Identifier of the question mark ("?")

val is_empty : t -> bool

val is_quest : t -> bool