Module Geneweb_db.Dutil

type name_index_data = int array array

Index for all kind of mix between person's names (first index inside names.inx)

type strings_of_fsname = int array array

Index for sub-strings of person's surame and first name (second and third index respectively inside names.inx)

val magic_GnWb0020 : string

Header for the base file (version 0020)

val magic_GnWb0021 : string

Header for the base file (version 0021)

val magic_GnWb0022 : string

Header for the base file (version 0022)

val magic_GnWb0023 : string

Header for the base file (version 0023)

val magic_GnWb0024 : string

Header for the latest version of base file

val table_size : int

Maximal size of hash table for name indexation (inside names.inx)

val compare_fnames_i : Dbdisk.base_data -> int -> int -> int

compare_fnames_i base i1 i2 compare two first names that have indexes i1 and i2 inside the base.

val compare_fnames : string -> string -> int

compare_fnames compare two first names.

val compare_snames_i : Dbdisk.base_data -> int -> int -> int

compare_snames_i base i1 i2 compare two surnames that have indexes i1 and i2 inside the base.

val compare_snames : Dbdisk.base_data -> string -> string -> int

compare_snames_i base s1 s2 compare two surnames according to the principe specified by Mutil.compare_after_particle.

val dsk_person_misc_names : Dbdisk.dsk_base -> Dbdisk.dsk_person -> (Dbdisk.dsk_person -> Dbdisk.dsk_title list) -> string list

dsk_person_misc_names base p nobtit computes various mix between all kind of names of a person's entry p from the database base. nobtit is used to return a title entries for passed in argument person.

val poi : Dbdisk.dsk_base -> int -> Dbdisk.dsk_person

poi base i returns person's entry with index i from base.

val sou : Dbdisk.dsk_base -> int -> string

poi base i returns string with index i from base.

val p_first_name : Dbdisk.dsk_base -> Dbdisk.dsk_person -> string

Returns person's first name from the given person's entry.

val p_surname : Dbdisk.dsk_base -> Dbdisk.dsk_person -> string

Returns person's surname from the given person's entry.

val output_value_no_sharing : Stdlib.out_channel -> _ -> unit

Output given value to the channel. Uses Marshall.to_channel with No_sharing flag.

val int_size : int

Size of integer value inside the Geneweb's binary files

module IntHT : sig ... end

Hastable that has unhashed int as a key.

val name_index : string -> int

name_index s Compute the index of crush_lowered version of s in an array of size table_size.