Module Geneweb_db.Outbase

val save_mem : bool Stdlib.ref

Flag that enables memory saving by calling gc sometimes

val set_ngram_thresholds : int -> int -> int -> unit

set_ngram_thresholds bi tri quad Configure n-gram indexing with minimum occurrence thresholds for bigrams, trigrams and quadrigrams. Call before rebuild to enable n-gram enrichment of names.inx index.

val output : Dbdisk.dsk_base -> unit

output base uses data section of the base to store database on the disk in the files:

base main file that stores all the arrays of the database

base.acc direct accesses to arrays inside base (list of offsets for every array)

names.inx 3 different name indexes :

For all kind of mix between person's names. Associate hash value of the name to the array of persons (index of its entry) containing the given name

sub-strings of surname. Associate hash value of the sub-string to the array of string indexes of the names that contains mentionned sub string.

sub-strings of first name. Same storage principe as for surname sub-strings.

names.acc direct accesses to arrays inside names.inx.

strings.inx strings index. Associate hash of the string to the index of its entry in the base's string array. Save also previus value in the case of collision of hash.

snames.inx ordered index for surnames. Associate index of surname to its offset in snames.dat.

snames.dat For a giving surname give associated list of perosons (index of its entry)

fnames.inx and fnames.dat same as for snames.inx and snames.dat but deals with first names

notes text file containing data base notes.

notes_d directory containing .txt for each extended page.

particles.txt text file with autorised name's particles.