Module Geneweb_search

module Analyze : sig ... end

module Automaton : sig ... end

A Levenshtein automaton for an alphabet A, a pattern p in A* and a maximal distance d is NFA that recognizes the language of all the words in A* that are at Levenshtein distance at most d of p.

module Flatset : sig ... end

module Heap : sig ... end

This module implements a simple imperative priority heap.

module Index : sig ... end

This module provides an implementation of reversed index based on trie data structures.

module Iterator : sig ... end

module Trie : sig ... end

module Word : sig ... end