Module Automaton.Make

Create a Levenshtein automaton for the alphabet W, the pattern pattern and the maximal Levenshtein distance max_dist.

Parameters

module W : Word.S

module _ : X with type word = W.t

Signature

type char_ = W.char_

type word = W.t

type state

State of the automaton.

val init : state

The initial state of the automaton.

val next : char_ -> state -> state

next c st returns the next state of the automaton after reading the character c.

val accept : state -> bool

accept st determines if the state st is accepted by the automaton.

val can_match : state -> bool

val recognize : word -> bool

recognize s determines if the string s is recognized by the automaton.