Module Index.Default

type t

type char_ = char

type word = string

type entry = string

val of_seq : (word * entry) Stdlib.Seq.t -> t

of_seq s creates a new inverted index from the sequence s. The sequence s is forced.

val mem : word -> t -> bool

mem w t checks if the word w is in the index t.

val search : word list -> t -> entry Stdlib.Seq.t

search ws t returns the sequence of entries in the index t which are associated with all the exact words ws.

val search_prefix : word list -> t -> entry Stdlib.Seq.t

search_prefix ps t returns the sequence of entries in the index t which are associated with all the prefix ps.

val fuzzy_search : max_dist:int -> word list -> t -> entry Stdlib.Seq.t

fuzzy_search ~max_dist ps t returns the sequence of entries in the index t which are associated with all the words matching each pattern of ps with a Levenstein distance limited to max_dist.