Module Gwdb_driver.Collection

Collections of elemetns

type 'a t

Collections are sets of elements you want to traverse.

val length : 'a t -> int

Return the number of elements of a colletion

val map : ('a -> 'b) -> 'a t -> 'b t

map fn c Return a collection corresponding to c where fn would have been applied to each of its elements.

val iter : ('a -> unit) -> 'a t -> unit

iter fn c Apply fn would have been applied to each elements of c.

val iteri : (int -> 'a -> unit) -> 'a t -> unit

iter fn c Apply fn i would have been applied to each elements of c where i is the index (starting with 0) of the element.

val fold : ?from:int -> ?until:int -> ('a -> 'b -> 'a) -> 'a -> 'b t -> 'a

fold fn acc c Combine each element of c into a single value using fn. fn first argument is the result computed so far as we traverse the collection, and second element is the current element being combined. acc is the starting combined value. Start at from-nth and finish with until-nth element (included).

val fold_until : ('a -> bool) -> ('a -> 'b -> 'a) -> 'a -> 'b t -> 'a

fold_until continue fn acc c Same as fold fn acc c, but computation stops as soon as continue is not satisfied by combined value anymore.

val iterator : 'a t -> unit -> 'a option

iterator c Return a function returning Some next_element when it is called, or None if you reached the end of the collection.