Module Gwdb_driver.Marker

Markers for elements inside Collection.t

type ('k, 'v) t

Markers are way to annotate (add extra information to) elements of a Collection.t.

val get : ('k, 'v) t -> 'k -> 'v

get marker key Return the annotation associated to key.

val set : ('k, 'v) t -> 'k -> 'v -> unit

set marker key value Set value as annotation associated to key.