Module Gwexport

type gwexport_charset =

| Ansel

| Ansi

| Ascii

| Utf8

type gwexport_opts = {

asc : int option;

ascdesc : int option;

base : (string * Gwdb.base) option;

censor : int;

charset : gwexport_charset;

desc : int option;

img_base_path : string;

keys : string list;

mem : bool;

no_notes : [ `nn | `nnn | `none ];

no_picture : bool;

oc : string * (string -> unit) * (unit -> unit);

parentship : bool;

picture_path : bool;

source : string option;

surnames : string list;

verbose : bool;

}

val default_opts : gwexport_opts

Default set of options

val speclist : gwexport_opts Stdlib.ref -> (Stdlib.Arg.key * Stdlib.Arg.spec * Stdlib.Arg.doc) list

Given a set of options, returns default command line arguments for selecting elements from a base. The output of this function is the first input of Arg.parse.

val anonfun : gwexport_opts Stdlib.ref -> Stdlib.Arg.anon_fun

anonfun opts = fun base_name -> ... Given a set of options `opts` where `!opts.base` is uninitialized, opens the dir `base_name` and initializes !opts.base with the base name. The output of this function is the second argument of Arg.parse.

val errmsg : Stdlib.Arg.usage_msg

Default error message. This is the third argument of Arg.parse.

val select : gwexport_opts -> Gwdb.iper list -> (Gwdb.iper -> bool) * (Gwdb.ifam -> bool)

select opts ips Return filters for iper and ifam to be used when exporting a (portion of a) base.