Module My_unix

val waitpid_noeintr : Unix.wait_flag list -> int -> int * Unix.process_status

Equivalent to Unix.waitpid, but this variant restart upon interruption by signals.

val accept_noeintr : ?cloexec:bool -> Unix.file_descr -> Unix.file_descr * Unix.sockaddr

Equivalent to Unix.accept, but this variant restart upon interruption by signals.