Module type Pqueue.OrderedType

The input signature of the functor Pqueue.Make. t is the type of the inserted elements. leq is a total ordering function over the elements. This is a two-argument function f returning True if the first argument is less or equal to the second one.

type t

val leq : t -> t -> bool