Module type Pqueue.S

Output signature for priority queue

type elt

Type of elementes contained in priority queues.

type t

Type of priority queues.

val empty : t

The empty queue.

val is_empty : t -> bool

is_empty q checks the emptiness of q.

val add : elt -> t -> t

add x h adds a new element x in heap h.

val take : t -> elt * t

take x removes the minimum element of x and returns it; raises Not_found when x is empty.

val union : t -> t -> t

union q1 q2 returns heap constructed by union of q1 q2