Module Wserver

type handler = (Unix.sockaddr * string list) -> string -> Adef.encoded_string -> unit

val start : ?addr:string -> port:int -> ?timeout:int -> max_pending_requests:int -> n_workers:int -> handler -> unit

Wserver.start ~secret_salt ?addr ~port ?timeout ~n_workers callback starts a HTTP 1.1 server that listens on the address addr and port port.

On Unix, worker jobs managed by n_workers workers have a time limit of timeout. If timeout is 0, there is no limit. This is the default.

The max_pending_requests argument specified the maximum number of pending requests that the server can store. If the queue is full, new requests are ignored until space becomes available.

When a client connects, callback is invoked with the arguments (addr, request) path query where:

addr is the client address,

request is the client request,

path is the path of the request,

query is the query content.

Listening on ports < 1024 may require root privileges.

val close_connection : unit -> unit

Closes the current socket

val printf : ('a, Stdlib.out_channel, unit) Stdlib.format -> 'a

Formatter printing in the out channel associated to the connected socket

val print_string : string -> unit

Prints a string in the out channel associated to the socket

val header : string -> unit

Prints a header; cannot be called if part of content part already has been sent

val wflush : unit -> unit

Flushes the content of the current socket

val string_of_status : Def.httpStatus -> string

HTTP status line (e.g. "200 OK", "404 Not Found").

val http : Def.httpStatus -> unit

Output.status conf answer sends the http header where answer represents the answer status.

val http_redirect_temporarily : string -> unit

Output.status conf_redirect url sends the http header where url represents the Location where the request needs to be redirected.

val get_request_and_content : char Stream.t -> string list * Adef.encoded_string

Returns the request from a stream read from a socket.

val wsocket : unit -> Unix.file_descr

Returns the last used socket

val woc : unit -> Stdlib.out_channel

Return the out_channel associated to the socket

val sock_in : string Stdlib.ref

Names of the files used in windows implementation to communicate http requests and html answers. Default "wserver.sin" and "wserver.sou". Can have relative or absolute paths.

val sock_out : string Stdlib.ref

val stop_server : string Stdlib.ref

Name of the file whose presence tells the server to stop (at least one request is necessary to unfreeze the server to make it check that this file exits. Default "STOP_SERVER". Can have relative or absolute path.

val cgi : bool Stdlib.ref

CGI (Common Gateway Interface) mode (default false).

module Pool : module type of Pool