Module Geneweb.GWPARAM

val nb_errors : int Stdlib.ref

val errors_undef : string list Stdlib.ref

val errors_other : string list Stdlib.ref

val set_vars : string list Stdlib.ref

val gwd_cmd : string Stdlib.ref

val cnt_dir : string Stdlib.ref

val sock_dir : string Stdlib.ref

val bases : string Stdlib.ref

val reorg : bool Stdlib.ref

set to true when base is in reorg format

val force : bool Stdlib.ref

force creation of database if already existing

val config_reorg : string -> string

val config_legacy : string -> string

type my_fun_2 = string -> string

type my_fun_3 = string -> string -> string

val config : my_fun_2 Stdlib.ref

val cnt_d : my_fun_2 Stdlib.ref

val adm_file : my_fun_2 Stdlib.ref

val src_d : my_fun_2 Stdlib.ref

val etc_d : my_fun_2 Stdlib.ref

val config_d : my_fun_2 Stdlib.ref

val lang_d : my_fun_3 Stdlib.ref

val bpath : my_fun_2 Stdlib.ref

val portraits_d : my_fun_2 Stdlib.ref

val images_d : my_fun_2 Stdlib.ref

val init : unit -> unit

Function called to initialize path functions based on the reorg flag.

val is_reorg_base : string -> bool

returns true if mybase.gwb/config/mybase.gwf exists

val test_reorg : string -> unit

set reorg to true if !reorg of is_reorg_base; then calls init

val check_base_exists : string -> unit

Check if database exists and exit if so (unless -f flag is set)

val create_base_and_config : string -> string

Create base directory and configuration.

module Default : sig ... end

module Legacy : sig ... end

val output_error : ?headers:string list -> ?content:Adef.safe_string -> Config.config -> Def.httpStatus -> unit

If ?content is not set, sends page content from /etc/<status-code>-<lang>.html. If the current lang is not available, use `en`

val is_semi_public : Geneweb_db.Driver.person -> bool

determines if the person has SemiPublic status

val split_key : string -> string * string * string

split a key of the form first_name.occ surname into its three components the .occ part may be absent. No spaces in first_name and surnames

val is_related : Config.config -> Geneweb_db.Driver.base -> Geneweb_db.Driver.person -> bool

determines if the person is a descendant a sibling or an ancestor of conf.userkey. conf.userkey is the person visiting the base the search for ancestors is limited to 3 generations

val has_date : Geneweb_db.Driver.base -> Geneweb_db.Driver.person -> bool

person has at least one date (birth, baptism, marriage, death, burial)

val p_auth : Config.config -> Geneweb_db.Driver.base -> Geneweb_db.Driver.person -> bool

Calculate the access rights to the person's information in according to his age. Returns (in the order of the tests) :

`true` if requester is wizard or friend or person is public

`true` if person has at least one title and public_if_title is set to yes in gwf config file

`false` if person is alive and private_years > 0

`true` if person is older (depending on the date of birth or baptism date) then privates_years

`false` if person is younger (depending on the date of birth or baptism date) then privates_years

`true` if person has been deceased for more than privates_years

`false` if person has been deceased for less than privates_years

`true` if person is between 80 and 120 years old and he is not beeing private and public_if_no_date is set to yes in gwf config file

`true` if person has been married for more than private_years

`false` otherwise

val p_auth_sp : Config.config -> Geneweb_db.Driver.base -> Geneweb_db.Driver.person -> bool

returns p_auth or true if both user and p are SemiPublic

val wrap_output : Config.config -> Adef.safe_string -> (unit -> unit) -> unit

Display in a very basic HTML doc, with no CSS or JavaScript.